requirements.txt
requests
flask
schedule
python-dotenv